refactor: common fn for mutable add
